+++ CHANGES.md
@@ -2,13 +2,14 @@
 
 * Added mappings for some iTunes Radio codes
 * Added support for non-dictionary list codes such as daap.browseartistlisting
+* Added --raw-codes option to dmapprint
 * When displaying the output of /content-codes dmapprint now includes a string representation of the code's type
 * Fixed handling of unsigned 8-bit values
 
 1.1.0 (2013-10-05)
 
 * Added `dmap_version()` and `dmap_version_string()`
-* Added --version argument to dmapprint
+* Added --version option to dmapprint
 * Added mappings for Digital Photo Access Protocol content codes
 * Added mappings for chapter-data and media-kind-listing content codes
 * Fixed NULL input parameter handling